requirements.txt
python-dotenv
streamlit
langchain_community
langchain-google-genai
faiss-cpu
langchain-groq
langchain


